Mods can now be installed locally

+++ src/r2mm/installing/ZipExtract.ts
@@ -8,23 +8,31 @@ import * as path from 'path';
 export default class ZipExtract {
 
     public static extractAndDelete(zipFolder: string, filename: string, outputFolderName: string, callback: (success: boolean) => void): ZipExtrationError | null {
+        return this.extractOnly(path.join(zipFolder, filename), path.join(zipFolder, outputFolderName), result => {
+            if (result) {
+                try {
+                    fs.removeSync(path.join(zipFolder, filename));
+                } catch(e) {
+                    const err: Error = e;
+                    return new FileWriteError(
+                        'Failed to delete file',
+                        err.message,
+                        null
+                    );
+                }
+            }
+            return result;
+        })
+    }
+
+    public static extractOnly(zip: string, outputFolder: string, callback: (success: boolean) => void): ZipExtrationError | null {
         try {
-            fs.createReadStream(path.join(zipFolder, filename))
-                .pipe(unzipper.Extract({ path: path.join(zipFolder, outputFolderName) })).promise().then(()=>{
-                    callback(true);
-                    try {
-                        fs.removeSync(path.join(zipFolder, filename));
-                    } catch(e) {
-                        const err: Error = e;
-                        return new FileWriteError(
-                            'Failed to delete file',
-                            err.message,
-                            null
-                        );
-                    }
-                }).catch(()=>{
-                    callback(false);
-                });
+            fs.createReadStream(zip)
+                .pipe(unzipper.Extract({ path: outputFolder })).promise().then(()=>{
+                callback(true);
+            }).catch(()=>{
+                callback(false);
+            });
         } catch(e) {
             const err: Error = e;
             return new ZipExtrationError(
